requirements.txt
# requirements.txt
Flask>=2.0
google-generativeai>=0.3
python-dotenv>=0.19